fixed the prof view
With the help of ahmed, made the prof view work correctly
