--- src/main/java/Pack/HomeController.java
@@ -1,14 +1,0 @@
-package Pack;
-
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-public class HomeController {
-
-    @GetMapping("/")
-    public String index() {
-        return "index";
-    }
-
-
-}
