--- src/main/resources/templates/AddressBook.html
@@ -1,11 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Printed AddressBook</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-<p th:each = "buddy: ${addresbook.getTopics()}"
-   th:text = "${buddy.toString()}"></p>
-</body>
-</html>
