+++ src/main/resources/templates/Display.html
@@ -2,26 +2,31 @@
 <html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
 <head>
     <meta charset="UTF-8">
-    <title>Address Book Contents</title>
+    <title>Prof Topic Contents</title>
 </head>
 <body>
-<h1>Your Address Book</h1>
+<h1>Your Topics</h1>
 <table>
     <thead>
     <tr>
-        <th> Name </th>
-        <th> Number </th>
-        <th> Address </th>
+        <th> Topic Title </th>
+        <th> Description </th>
+        <th> Program Restriction </th>
+        <th> Max Students Allowed </th>
+
     </tr>
     </thead>
     <tbody>
-    <tr th:if="${buddies.empty}">
-        <td colspan="3"> Empty Address Book </td>
+    <tr th:if="${topic.empty}">
+        <td colspan="3"> You do not have any current topics </td>
     </tr>
-    <tr th:each="buddy : ${buddies}">
-        <td><span th:text="${buddy.name}"> Name </span></td>
-        <td><span th:text="${buddy.number}"> Number </span></td>
-        <td><span th:text="${buddy.address}"> Address </span></td>
+    <tr th:each="topic : ${topic}">
+        <td><span th:text="${topic.topicName}"> Topic Title </span></td>
+        <td><span th:text="${topic.description}"> Description </span></td>
+        <td><span th:text="${topic.restrictions}"> Program Restriction </span></td>
+        <td><span th:text="${topic.max_Students}"> Max Students Allowed </span></td>
+
+
     </tr>
     </tbody>
 </table>
