+++ src/main/resources/templates/index.html
@@ -15,7 +15,7 @@
     margin-right: auto;">
         <button class="btn btn-outline-success" type="submit" formaction="/add">Professor</button>
         <button class="btn btn-outline-danger" type="submit" formaction="/add">Coordinator</button>
-        <button class="btn btn-outline-primary" type="submit" formaction="/student">Student</button>
+        <button class="btn btn-outline-primary" type="submit" formaction="/students">Student</button>
     </form>
 </nav>
 <!-- Optional JavaScript -->
